livekit-agent/requirements.txt
# LiveKit Agents Dependencies - September 2025
livekit>=1.0.13
livekit-agents>=1.2.9
livekit-plugins-google>=1.2.9
livekit-plugins-silero>=1.2.9
livekit-plugins-deepgram>=1.2.9
livekit-plugins-turn-detector>=1.0.0
python-dotenv>=1.0.0
supabase>=2.9.0
aiohttp>=3.10.0
numpy>=2.1.0
google-generativeai>=0.8.0